Rename css folder to scss
